requirements.txt
aniso8601==7.0.0
appdirs==1.4.3
asn1crypto==0.24.0
astroid==2.2.5
attrs==19.1.0
Automat==0.7.0
Babel==2.7.0
beautifulsoup4==4.8.0
black==19.3b0
certifi==2019.6.16
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.0.3
entrypoints==0.3
flake8==3.7.8
gunicorn==19.9.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.1
lxml==4.4.0
Mako==1.1.0
MarkupSafe==1.1.1
marshmallow==2.20.0
mccabe==0.6.1
parsel==1.5.1
pyasn1==0.4.6
pyasn1-modules==0.2.6
pycodestyle==2.5.0
pycparser==2.19
PyDispatcher==2.0.5
pyflakes==2.1.1
PyHamcrest==1.9.0
pylint==2.3.1
pyOpenSSL==19.0.0
python-dateutil==2.8.0
python-dotenv==0.10.3
python-editor==1.0.4
pytz==2019.2
queuelib==1.5.0
requests==2.22.0
Scrapy3==1.0.1
service-identity==18.1.0
six==1.12.0
soupsieve==1.9.2
toml==0.10.0
Twisted==19.2.1
typed-ast==1.4.0
urllib3==1.25.3
urlparse3==1.1
w3lib==1.20.0
Werkzeug==0.15.5
wrapt==1.11.2
zope.interface==4.6.0
